requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.14.0
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.13.0
async-timeout==3.0.1
attrs==24.2.0
backoff==1.11.1
beautifulsoup4==4.13.5
bidict==0.23.1
blinker==1.9.0
blis==1.1.0
boolean.py==5.0
CacheControl==0.14.4
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.6
cyclonedx-python-lib==11.7.0
cymem==2.0.10
DateTime==5.5
defusedxml==0.7.1
discord.py==2.4.0
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.135.3
filelock==3.25.2
Flask==3.1.3
frozenlist==1.5.0
google-api-core==2.30.2
google-api-python-client==2.156.0
google-auth==2.49.1
google-auth-httplib2==0.2.0
google-genai==1.70.0
googleapis-common-protos==1.74.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
license-expression==30.4.4
Mako==1.3.8
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.2
multidict==6.1.0
murmurhash==1.0.11
nltk==3.9.4
numpy==2.2.0
obsws-python==1.7.0
openai==2.14.0
packageurl-python==0.17.6
packaging==24.2
pefile==2023.2.7
pillow==12.2.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.9.4
preshed==3.0.9
propcache==0.2.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==6.1.1
py-serializable==2.1.0
pyasn1==0.6.3
pyasn1_modules==0.4.1
pycparser==3.0
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.20.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pyparsing==3.2.0
python-dotenv==1.0.1
python-engineio==4.11.1
python-socketio==5.16.1
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.33.1
rich==13.9.4
rsa==4.9
schedule==1.2.2
shellingham==1.5.4
simple-websocket==1.1.0
simpleaudio==1.0.4
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spotify==0.10.2
spotipy==2.26.0
SQLAlchemy==2.0.36
srsly==2.5.0
starlette==1.0.0
tenacity==9.1.4
textblob==0.18.0.post0
thinc==8.3.3
tomli==2.4.1
tomli_w==1.2.0
tqdm==4.67.1
twitchio==2.10.0
typer==0.15.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.1.1
urllib3==2.6.3
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.8
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zope.interface==7.2
